requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.6
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
idna==2.9
Pillow==7.1.2
pytz==2020.1
rajaongkir==1.1.2
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
